+++ test/unittests/commands/validate_data/conftest.py
@@ -1,4 +1,5 @@
 from test.unittests.commands.validate_data.fixtures_data_deserialised import *  # noqa: F403
 from test.unittests.commands.validate_data.fixtures_data_rdf import *  # noqa: F403
 from test.unittests.commands.validate_data.fixtures_onto import *  # noqa: F403
+from test.unittests.commands.validate_data.fixtures_validation_result import *  # noqa: F403
 from test.unittests.commands.validate_data.fixtures_xml_data import *  # noqa: F403
